process/pyproject.toml
[project]
name = "process"
version = "0.1.0"
description = "Process Microservice"
requires-python = ">=3.11"
dependencies = [
    # Core web framework dependencies
    "fastapi>=0.68.0",
    "uvicorn>=0.15.0",
    # process特有依赖
    "mysqlclient>=2.2.7",
    "danloo-common",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
test = [
    # process特有的test依赖（如果有的话）
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.hatch.metadata]
allow-direct-references = true

# Removed workspace reference - using direct path installation in Docker

[tool.pytest.ini_options]
addopts = "--cov=services --cov=models --cov=controllers --tb=short"
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
asyncio_mode = "auto"

[tool.coverage.run]
source = ["services", "models", "controllers"]
omit = [
    "tests/*",
    "venv/*",
    "env/*"
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
]

[dependency-groups]
test = [
    # process特有的test依赖（如果有的话）
]
